Created a DAL abstract class, then separated databases into their own files that inherit this class.
